+++ controllers/food.js
@@ -40,74 +40,7 @@ const getFoodDataById = async (req, res) => {
 
 const createFoodData = async (req, res) => {
   try {
-    let failMessage = '';
-    const food = {
-      name: req.body.name,
-      description: req.body.description,
-      recipe: req.body.recipe,
-      weight: req.body.weight,
-      stack: req.body.stack,
-      maxHealth: req.body.maxHealth,
-      maxStamina: req.body.maxStamina,
-      maxEitr: req.body.maxEitr,
-      duration: req.body.duration,
-      healing: req.body.healing
-    };
-    if (typeof food.name != 'string') {
-      failMessage += 'To create Food data, enter a name string.\n';
-    }
-    if (typeof food.description != 'string') {
-      failMessage += 'To create Food data, enter a description string.\n';
-    }
-    if (typeof food.recipe != 'string') {
-      failMessage += 'To create Food data, enter a recipe string.\n';
-    }
-    if (typeof food.weight != 'number') {
-      failMessage += 'To create Food data, enter a numeric weight amount.\n';
-    }
-    if (typeof food.stack != 'number') {
-      failMessage += 'To create Food data, enter a numeric stack amount.\n';
-    }
-    if (typeof food.maxHealth != 'number') {
-      failMessage += 'To create Food data, enter a numeric maxHealth amount.\n';
-    }
-    if (typeof food.maxStamina != 'number') {
-      failMessage += 'To create Food data, enter a numeric maxStamina amount.\n';
-    }
-    if (typeof food.maxEitr != 'number') {
-      failMessage += 'To create Food data, enter a numeric maxEitr amount.\n';
-    }
-    if (typeof food.duration != 'string') {
-      failMessage += 'To create Food data, enter a duration string.\n';
-    }
-    if (typeof food.healing != 'string') {
-      failMessage += 'To create Food data, enter a healing string.';
-    }
-    if (failMessage != '') {
-      res.status(400);
-      res.send(failMessage);
-    } else {
-      const responce = await mongodb.getDb().db('valheim').collection('food').insertOne(food);
-      if (responce.acknowledged) {
-        res.status(201).json(responce);
-      } else {
-        res
-          .status(500)
-          .json(
-            responce.error || 'Something went wrong while creating the food data. Try again later.'
-          );
-      }
-    }
-  } catch (err) {
-    res.status(500).json('Something went wrong while creating the food data. Try again later.');
-  }
-};
-
-const updateFoodData = async (req, res) => {
-  try {
-    if (!ObjectId.isValid(req.params.id)) {
-      res.status(400).json('Id must be alphanumeric, 24 characters long.');
-    } else {
+    if (req.oidc.isAuthenticated()) {
       let failMessage = '';
       const food = {
         name: req.body.name,
@@ -122,72 +55,148 @@ const updateFoodData = async (req, res) => {
         healing: req.body.healing
       };
       if (typeof food.name != 'string') {
-        failMessage += 'To update Food data, enter a name string.\n';
+        failMessage += 'To create Food data, enter a name string.\n';
       }
       if (typeof food.description != 'string') {
-        failMessage += 'To update Food data, enter a description string.\n';
+        failMessage += 'To create Food data, enter a description string.\n';
       }
       if (typeof food.recipe != 'string') {
-        failMessage += 'To update Food data, enter a recipe string.\n';
+        failMessage += 'To create Food data, enter a recipe string.\n';
       }
       if (typeof food.weight != 'number') {
-        failMessage += 'To update Food data, enter a numeric weight amount.\n';
+        failMessage += 'To create Food data, enter a numeric weight amount.\n';
       }
       if (typeof food.stack != 'number') {
-        failMessage += 'To update Food data, enter a numeric stack amount.\n';
+        failMessage += 'To create Food data, enter a numeric stack amount.\n';
       }
       if (typeof food.maxHealth != 'number') {
-        failMessage += 'To update Food data, enter a numeric maxHealth amount.\n';
+        failMessage += 'To create Food data, enter a numeric maxHealth amount.\n';
       }
       if (typeof food.maxStamina != 'number') {
-        failMessage += 'To update Food data, enter a numeric maxStamina amount.\n';
+        failMessage += 'To create Food data, enter a numeric maxStamina amount.\n';
       }
       if (typeof food.maxEitr != 'number') {
-        failMessage += 'To update Food data, enter a numeric maxEitr amount.\n';
+        failMessage += 'To create Food data, enter a numeric maxEitr amount.\n';
       }
       if (typeof food.duration != 'string') {
-        failMessage += 'To update Food data, enter a duration string.\n';
+        failMessage += 'To create Food data, enter a duration string.\n';
       }
       if (typeof food.healing != 'string') {
-        failMessage += 'To update Food data, enter a healing string.';
+        failMessage += 'To create Food data, enter a healing string.';
       }
       if (failMessage != '') {
         res.status(400);
         res.send(failMessage);
       } else {
-        const userId = new ObjectId(req.params.id);
-        const responce = await mongodb
-          .getDb()
-          .db('valheim')
-          .collection('food')
-          .updateOne(
-            { _id: userId },
-            {
-              $set: {
-                name: food.name,
-                description: food.description,
-                recipe: food.recipe,
-                weight: food.weight,
-                stack: food.stack,
-                maxHealth: food.maxHealth,
-                maxStamina: food.maxStamina,
-                maxEitr: food.maxEitr,
-                duration: food.duration,
-                healing: food.healing
-              }
-            }
-          );
-        if (responce.modifiedCount > 0) {
-          res.status(204).send();
+        const responce = await mongodb.getDb().db('valheim').collection('food').insertOne(food);
+        if (responce.acknowledged) {
+          res.status(201).json(responce);
         } else {
           res
             .status(500)
             .json(
               responce.error ||
-                'Something went wrong while updating the food data. Try again later.'
+                'Something went wrong while creating the food data. Try again later.'
             );
         }
       }
+    } else {
+      res.status(401).json('You must login to run this request.');
+    }
+  } catch (err) {
+    res.status(500).json('Something went wrong while creating the food data. Try again later.');
+  }
+};
+
+const updateFoodData = async (req, res) => {
+  try {
+    if (req.oidc.isAuthenticated()) {
+      if (!ObjectId.isValid(req.params.id)) {
+        res.status(400).json('Id must be alphanumeric, 24 characters long.');
+      } else {
+        let failMessage = '';
+        const food = {
+          name: req.body.name,
+          description: req.body.description,
+          recipe: req.body.recipe,
+          weight: req.body.weight,
+          stack: req.body.stack,
+          maxHealth: req.body.maxHealth,
+          maxStamina: req.body.maxStamina,
+          maxEitr: req.body.maxEitr,
+          duration: req.body.duration,
+          healing: req.body.healing
+        };
+        if (typeof food.name != 'string') {
+          failMessage += 'To update Food data, enter a name string.\n';
+        }
+        if (typeof food.description != 'string') {
+          failMessage += 'To update Food data, enter a description string.\n';
+        }
+        if (typeof food.recipe != 'string') {
+          failMessage += 'To update Food data, enter a recipe string.\n';
+        }
+        if (typeof food.weight != 'number') {
+          failMessage += 'To update Food data, enter a numeric weight amount.\n';
+        }
+        if (typeof food.stack != 'number') {
+          failMessage += 'To update Food data, enter a numeric stack amount.\n';
+        }
+        if (typeof food.maxHealth != 'number') {
+          failMessage += 'To update Food data, enter a numeric maxHealth amount.\n';
+        }
+        if (typeof food.maxStamina != 'number') {
+          failMessage += 'To update Food data, enter a numeric maxStamina amount.\n';
+        }
+        if (typeof food.maxEitr != 'number') {
+          failMessage += 'To update Food data, enter a numeric maxEitr amount.\n';
+        }
+        if (typeof food.duration != 'string') {
+          failMessage += 'To update Food data, enter a duration string.\n';
+        }
+        if (typeof food.healing != 'string') {
+          failMessage += 'To update Food data, enter a healing string.';
+        }
+        if (failMessage != '') {
+          res.status(400);
+          res.send(failMessage);
+        } else {
+          const userId = new ObjectId(req.params.id);
+          const responce = await mongodb
+            .getDb()
+            .db('valheim')
+            .collection('food')
+            .updateOne(
+              { _id: userId },
+              {
+                $set: {
+                  name: food.name,
+                  description: food.description,
+                  recipe: food.recipe,
+                  weight: food.weight,
+                  stack: food.stack,
+                  maxHealth: food.maxHealth,
+                  maxStamina: food.maxStamina,
+                  maxEitr: food.maxEitr,
+                  duration: food.duration,
+                  healing: food.healing
+                }
+              }
+            );
+          if (responce.modifiedCount > 0) {
+            res.status(204).send();
+          } else {
+            res
+              .status(500)
+              .json(
+                responce.error ||
+                  'Something went wrong while updating the food data. Try again later.'
+              );
+          }
+        }
+      }
+    } else {
+      res.status(401).json('You must login to run this request.');
     }
   } catch (err) {
     res.status(500).json(err);
@@ -196,24 +205,29 @@ const updateFoodData = async (req, res) => {
 
 const deleteFoodData = async (req, res) => {
   try {
-    if (!ObjectId.isValid(req.params.id)) {
-      res.status(400).json('Id must be alphanumeric, 24 characters long.');
-    } else {
-      const userId = new ObjectId(req.params.id);
-      const responce = await mongodb
-        .getDb()
-        .db('valheim')
-        .collection('food')
-        .deleteOne({ _id: userId }, true);
-      if (responce.deletedCount > 0) {
-        res.status(200).send(`Food data with id ${userId} was deleted sucessfully.`);
+    if (req.oidc.isAuthenticated()) {
+      if (!ObjectId.isValid(req.params.id)) {
+        res.status(400).json('Id must be alphanumeric, 24 characters long.');
       } else {
-        res
-          .status(500)
-          .json(
-            responce.error || 'Something went wrong while deleting the food data. Try again later.'
-          );
+        const userId = new ObjectId(req.params.id);
+        const responce = await mongodb
+          .getDb()
+          .db('valheim')
+          .collection('food')
+          .deleteOne({ _id: userId }, true);
+        if (responce.deletedCount > 0) {
+          res.status(200).send(`Food data with id ${userId} was deleted sucessfully.`);
+        } else {
+          res
+            .status(500)
+            .json(
+              responce.error ||
+                'Something went wrong while deleting the food data. Try again later.'
+            );
+        }
       }
+    } else {
+      res.status(401).json('You must login to run this request.');
     }
   } catch (err) {
     res.status(500).json(err);
